feat: Extract fab menu to a dedicated component
Current implementation displays Fab menu on all application pages

We wan't to have a better control on when to display it or not

So it has been extracted to a dedicated component and it is now
displayed from specific pages instead of from App component
