+++ src/components/App.jsx
@@ -11,31 +11,20 @@ import useBreakpoints from 'cozy-ui/transpiled/react/hooks/useBreakpoints'
 import MuiCozyTheme from 'cozy-ui/transpiled/react/MuiCozyTheme'
 import Typography from 'cozy-ui/transpiled/react/Typography'
 import Alerter from 'cozy-ui/transpiled/react/Alerter'
-import Fab from 'cozy-ui/transpiled/react/Fab'
-import Icon from 'cozy-ui/transpiled/react/Icon'
-import PlusIcon from 'cozy-ui/transpiled/react/Icons/Plus'
 
 import HomeWrapper from 'src/components/HomeWrapper/HomeWrapper'
 import PapersList from 'src/components/Papers/PapersList'
 import FileViewerWithQuery from 'src/components/Viewer/FileViewerWithQuery'
 import { ModalStack } from 'src/components/Contexts/ModalProvider'
-import PlaceholderThemesList from 'src/components/Placeholders/PlaceholderThemesList'
-import { usePlaceholderModal } from 'src/components/Hooks/usePlaceholderModal'
 import { useStepperDialog } from 'src/components/Hooks/useStepperDialog'
 import StepperDialogWrapper from 'src/components/StepperDialogWrapper/StepperDialogWrapper'
 import { FormDataProvider } from 'src/components/Contexts/FormDataProvider'
-
-const styleFab = { position: 'fixed', zIndex: 10 }
 
 export const App = () => {
   const client = useClient()
   const { t } = useI18n()
   const { isMobile } = useBreakpoints()
   const { BarCenter } = cozy.bar
-  const {
-    showPlaceholderThemesList,
-    setShowPlaceholderThemesList
-  } = usePlaceholderModal()
 
   const { isStepperDialogOpen } = useStepperDialog()
 
@@ -69,21 +58,6 @@ export const App = () => {
                 <Redirect from="*" to="/" />
               </Switch>
 
-              <Fab
-                color="primary"
-                aria-label={t('Home.Fab.ariaLabel')}
-                style={styleFab}
-                className="u-bottom-m u-right-m"
-                onClick={() => setShowPlaceholderThemesList(true)}
-              >
-                <Icon icon={PlusIcon} />
-              </Fab>
-              {showPlaceholderThemesList && (
-                <PlaceholderThemesList
-                  title={t('PlaceholdersList.title', { name: '' })}
-                  onClose={() => setShowPlaceholderThemesList(false)}
-                />
-              )}
               {isStepperDialogOpen && (
                 <FormDataProvider>
                   <StepperDialogWrapper />
